+++ internal/meraki/devices/switch/port/resource.go
@@ -22,7 +22,7 @@ var (
 	_ resource.ResourceWithImportState = &Resource{} // Interface for resources with import state functionality
 )
 
-func NewDevicesSwitchPortResource() resource.Resource {
+func NewResource() resource.Resource {
 	return &Resource{}
 }
