+++ internal/meraki/networks/devices/claim/import.go
@@ -1,0 +1,14 @@
+package claim
+
+import (
+	"context"
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+func (r *Resource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+	resource.ImportStatePassthroughID(ctx, path.Root("network_id"), req, resp)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+}
