+++ internal/meraki/networks/devices/claim/resource_test.go
@@ -1,4 +1,4 @@
-package networks_test
+package claim_test
 
 import (
 	"fmt"
